requirements.txt
celery
flask
flask-cors
gunicorn
python-dotenv
openai
python-dotenv
redis